개발 완료 - 버그 약간 수정 (상품, 매장 삭제는 참조 엔티티 관련 문제로 기능을 중단함)

+++ src/main/java/com/knotted/service/admin/AdminBoardService.java
@@ -49,7 +49,9 @@ public class AdminBoardService {
         BoardImage boardImage = adminBoardImageRepository.findByBoardId(board.getId());
 
         // 게시글 이미지 파일 및 DB 먼저 제거
-        adminBoardImageService.deleteBoardImage(boardImage);
+        if(boardImage != null){ // 있으면 제거
+            adminBoardImageService.deleteBoardImage(boardImage);
+        }
 
         // 정상적으로 됐으면 게시글 제거
         adminBoardRepository.delete(board);
